+++ prost-reflect-tests/build.rs
@@ -64,6 +64,7 @@ fn main() -> io::Result<()> {
                 "src/test2.proto",
                 "src/desc.proto",
                 "src/desc_no_package.proto",
+                "src/import_any.proto",
             ],
             &["src/"],
         )?;