+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for JSON mapping of `google.protobuf.Any`.
+
 ## [0.3.2] - 2021-12-29
 
 ### Added